+++ sdk/formrecognizer/Azure.AI.FormRecognizer/src/DocumentAnalysisModelFactory.cs
@@ -14,12 +14,12 @@ namespace Azure.AI.FormRecognizer.DocumentAnalysis
     public static class DocumentAnalysisModelFactory
     {
         /// <summary> Initializes a new instance of ResourceDetails. </summary>
-        /// <param name="documentModelCount"> Number of custom models in the current resource. </param>
-        /// <param name="documentModelLimit"> Maximum number of custom models supported in the current resource. </param>
+        /// <param name="customDocumentModelCount"> Number of custom models in the current resource. </param>
+        /// <param name="customDocumentModelLimit"> Maximum number of custom models supported in the current resource. </param>
         /// <returns> A new <see cref="DocumentAnalysis.ResourceDetails"/> instance for mocking. </returns>
-        public static ResourceDetails ResourceDetails(int documentModelCount = default, int documentModelLimit = default)
+        public static ResourceDetails ResourceDetails(int customDocumentModelCount = default, int customDocumentModelLimit = default)
         {
-            return new ResourceDetails(documentModelCount, documentModelLimit);
+            return new ResourceDetails(customDocumentModelCount, customDocumentModelLimit);
         }
 
         /// <summary> Initializes a new instance of AddressValue. </summary>